Move some store types out into own file to reduce circular dependencies.
