+++ client/src/stores/activityStore.ts
@@ -1,7 +1,6 @@
 /**
  * Stores the Activity Bar state
  */
-import { type IconDefinition } from "@fortawesome/fontawesome-svg-core";
 import { useDebounceFn, watchImmediate } from "@vueuse/core";
 import { computed, type Ref, ref, set } from "vue";
 
@@ -10,37 +9,8 @@ import { useUserLocalStorage } from "@/composables/userLocalStorage";
 import { ensureDefined } from "@/utils/assertions";
 
 import { defaultActivities } from "./activitySetup";
+import type { Activity } from "./activityStoreTypes";
 import { defineScopedStore } from "./scopedStore";
-
-export type ActivityVariant = "primary" | "danger" | "disabled";
-
-export interface Activity {
-    // determine wether an anonymous user can access this activity
-    anonymous?: boolean;
-    // description of the activity
-    description: string;
-    // unique identifier
-    id: string;
-    // icon to be displayed in activity bar
-    icon: IconDefinition;
-    // indicate if this activity can be modified and/or deleted
-    mutable?: boolean;
-    // indicate wether this activity can be disabled by the user
-    optional?: boolean;
-    // specifiy wether this activity utilizes the side panel
-    panel?: boolean;
-    // title to be displayed in the activity bar
-    title: string;
-    // route to be executed upon selecting the activity
-    to?: string | null;
-    // tooltip to be displayed when hovering above the icon
-    tooltip: string;
-    // indicate wether the activity should be visible by default
-    visible?: boolean;
-    // if activity should cause a click event
-    click?: true;
-    variant?: ActivityVariant;
-}
 
 export interface ActivityMeta {
     disabled: boolean;
