+++ client/src/stores/workflowConnectionStore.test.ts
@@ -1,13 +1,8 @@
 import { createPinia, setActivePinia } from "pinia";
 
-import {
-    type Connection,
-    getTerminalId,
-    type InputTerminal,
-    type OutputTerminal,
-    useConnectionStore,
-} from "@/stores/workflowConnectionStore";
+import { getTerminalId, useConnectionStore } from "@/stores/workflowConnectionStore";
 import { type NewStep, useWorkflowStepStore } from "@/stores/workflowStepStore";
+import type { Connection, InputTerminal, OutputTerminal } from "@/stores/workflowStoreTypes";
 
 const workflowStepZero: NewStep = {
     input_connections: {},
