+++ client/src/stores/workflowStepStore.ts
@@ -1,11 +1,12 @@
 import { computed, del, ref, set } from "vue";
 
 import { type CollectionTypeDescriptor } from "@/components/Workflow/Editor/modules/collectionTypeDescription";
-import { type Connection, getConnectionId, useConnectionStore } from "@/stores/workflowConnectionStore";
+import { getConnectionId, useConnectionStore } from "@/stores/workflowConnectionStore";
 import { assertDefined } from "@/utils/assertions";
 
 import { defineScopedStore } from "./scopedStore";
 import { useWorkflowStateStore } from "./workflowEditorStateStore";
+import type { Connection } from "./workflowStoreTypes";
 
 interface StepPosition {
     top: number;
